Add files via upload
These are the individual photos in the Brain trust section. There are 3 versions of each person's photo (1. normal, 2. darkened, 3. darkened with text). 

I know you only needed 1 and 3, but I guess the designer provided 2 just to be safe.

We are going to send a layout with explicit measurements for the distance between the boxes, the names, the rows, etc...